+++ R/data-typical-data-large.R
@@ -1,8 +1,8 @@
 #' A small toy dataset of imaginary people
 #'
 #' A wider dataset than `typical_data` containing information about some
-#'   randomly generated people, created using the excellent `wakefield` package.
-#'   It is created as deliberately odd / eclectic dataset.
+#'   randomly generated people, created using the excellent `wakefield`
+#'   package. It is created as deliberately odd / eclectic dataset.
 #'
 #' @format A data frame with 300 rows and 49 variables:
 #' \describe{
@@ -35,8 +35,7 @@
 #'   \item{Level}{levels between 1 and 4}
 #'   \item{Likert}{likert response - "strongly agree", "agree", and so on}
 #'   \item{Lorem_Ipsum}{lorem ipsum text}
-#'   \item{Marital}{marital status- "married", "divorced", "widowed",
-#'     "separated", etc}
+#'   \item{Marital}{marital status- "married", "divorced", "widowed", "separated", etc}
 #'   \item{Military}{miliary branch they are in}
 #'   \item{Month}{their favorite month}
 #'   \item{Name}{their name}
@@ -57,5 +56,4 @@
 #'   \item{Year}{significant year to that individuals}
 #'   \item{Zip}{a zip code they have visited}
 #' }
-#'
 "typical_data_large"