+++ R/vis-miss.R
@@ -9,32 +9,55 @@
 #'
 #' @param cluster logical. TRUE specifies that you want to use hierarchical
 #'   clustering (mcquitty method) to arrange rows according to missingness.
-#'   FALSE specifies that you want to leave it as is.
+#'   FALSE specifies that you want to leave it as is. Default value is FALSE.
 #'
-#' @param sort_miss logical. TRUE arranges the columns in order of missingness
+#' @param sort_miss logical. TRUE arranges the columns in order of missingness.
+#'   Default value is FALSE.
 #'
 #' @param show_perc logical. TRUE now adds in the \% of missing/complete data
 #'   in the whole dataset into the legend. Default value is TRUE.
 #'
 #'@param show_perc_col logical. TRUE adds in the \% missing data in a given
-#'  column into the x axis. Can be disabled with FALSE
+#'  column into the x axis. Can be disabled with FALSE. Default value is TRUE.
 #'
-#' @param warn_large_data logical default is TRUE
+#' @param warn_large_data logical - warn if there is large data? Default is TRUE
+#'   see note for more details
 #'
-#' @param large_data_size integer default is 900000, this can be changed.
+#' @param large_data_size integer default is 900000, this can be changed. See
+#'   note for more details
 #'
 #' @return `ggplot2` object displaying the position of missing values in the
 #'   dataframe, and the percentage of values missing and present.
 #'
-#' @seealso [vis_dat()]
+#' @seealso [vis_dat()] [vis_guess()] [vis_expect()] [vis_cor()] [vis_compare()]
+#'
+#' @note Some datasets might be too large to plot, sometimes creating a blank
+#'   plot - if this happens, I would recommend downsampling the data, either
+#'   looking at the first 1,000 rows or by taking a random sample. This means
+#'   that you won't get the same "look" at the data, but it is better than
+#'   a blank plot! See example code for suggestions on doing this.
 #'
 #' @examples
 #'
 #' vis_miss(airquality)
 #'
+#' \dontrun{
 #' vis_miss(airquality, cluster = TRUE)
 #'
 #' vis_miss(airquality, sort_miss = TRUE)
+#'
+#' # if you have a large dataset, you might want to try downsampling:
+#' library(nycflight13)
+#' library(dplyr)
+#' flights %>%
+#'   sample_n(1000) %>%
+#'   vis_miss()
+#'
+#' flights %>%
+#'   slice(1:1000) %>%
+#'   vis_miss()
+#'
+#' }
 #'
 #' @export
 vis_miss <- function(x,
@@ -80,7 +103,7 @@ vis_miss <- function(x,
     # get the names of those columns
     col_order_index <- names(x)[na_sort]
 
-  } else {
+  } else if (!sort_miss) {
 
     col_order_index <- names(x)
 
@@ -88,10 +111,10 @@ vis_miss <- function(x,
 
   # Arranged data by dendrogram order index
 
-    # gather the variables together for plotting
-    # here we now have a column of the row number (row),
-    # then the variable(variables),
-    # then the contents of that variable (value)
+  # gather the variables together for plotting
+  # here we now have a column of the row number (row),
+  # then the variable(variables),
+  # then the contents of that variable (value)
   dat_pre_vis <- as.data.frame(x.na[row_order_index , ])
 
   d <- dat_pre_vis %>%
@@ -117,7 +140,6 @@ vis_miss <- function(x,
     p_pres_lab <- "Present"
 
   }
-
   # then we plot it
   vis_miss_plot <- vis_create_(d) +
     ggplot2::scale_fill_manual(name = "",
@@ -126,10 +148,41 @@ vis_miss <- function(x,
                                labels = c(p_pres_lab,
                                           p_miss_lab)) +
     ggplot2::guides(fill = ggplot2::guide_legend(reverse = TRUE)) +
-    ggplot2::theme(legend.position = "bottom")
+    ggplot2::theme(legend.position = "bottom") +
+    # fix up the location of the text
+    ggplot2::theme(axis.text.x = ggplot2::element_text(hjust = 0))
 
   # add the missingness column labels
-  if (show_perc_col){
+
+  # if there is only one colummn you don't need to sort the columns
+  # this is perhaps a bit of a hacky way around, but I can't see another
+  # way around it. Related issue: https://github.com/ropensci/visdat/issues/72
+  if (ncol(x) == 1) {
+
+    if (show_perc_col) {
+      return(
+        # print(
+      vis_miss_plot +
+        ggplot2::scale_x_discrete(position = "top",
+                                  labels = label_col_missing_pct(
+                                    x,
+                                    col_order_index)
+        )
+      # )
+      )
+    } else if (!show_perc_col) {
+      return(
+        # print(
+      vis_miss_plot +
+        ggplot2::scale_x_discrete(position = "top",
+                                  labels = col_order_index)
+      # )
+      )
+    }
+
+  }
+
+  if (show_perc_col) {
 
     # flip the axes, add the info about limits
     vis_miss_plot +
@@ -138,18 +191,17 @@ vis_miss <- function(x,
                                 labels = label_col_missing_pct(
                                   x,
                                   col_order_index)
-                                ) +
-      # fix up the location of the text
-      ggplot2::theme(axis.text.x = ggplot2::element_text(hjust = 0))
+      )
+
   } else {
     vis_miss_plot +
-    ggplot2::scale_x_discrete(position = "top",
-                              limits = col_order_index)
+      ggplot2::scale_x_discrete(position = "top",
+                                limits = col_order_index)
   }
 
-      # guides(fill = guide_legend(title = "Type"))
+  # guides(fill = guide_legend(title = "Type"))
   # Thanks to
-# http://www.markhneedham.com/blog/2015/02/27/rggplot-controlling-x-axis-order/
+  # http://www.markhneedham.com/blog/2015/02/27/rggplot-controlling-x-axis-order/
   # For the tip on using scale_x_discrete
 
-} # end of function
+  } # end of function