+++ R/visdat-package.r
@@ -10,13 +10,26 @@
 #' \itemize{
 #'   \item [vis_dat()]
 #'   \item [vis_miss()]
+#'   \item [vis_guess()]
+#'   \item [vis_compare()]
+#'   \item [vis_expect()]
 #' }
 #'
 #' Learn more about visdat at \url{www.njtierney.com/visdat/articles/using_visdat.html}
 #' @name visdat
 #' @docType package
 #' @importFrom magrittr %>%
+#' @importFrom stats cor
+#' @importFrom stats setNames
 NULL
 
 if(getRversion() >= "2.15.1")  utils::globalVariables(c("."))
-globalVariables(c("valueGuess"))
+globalVariables(c("valueGuess",
+                  "valueType",
+                  "row_1",
+                  "row_2",
+                  "value",
+                  "cor",
+                  "setNames",
+                  "rowname",
+                  "n"))