+++ inst/doc/using_visdat.R
@@ -10,13 +10,14 @@ head(iris)
 
 
 ## ----glimpse-------------------------------------------------------------
-
-dplyr::glimpse(iris)
+library(dplyr)
+glimpse(iris)
 
 
 ## ----visdat-glimpse------------------------------------------------------
 library(visdat)
-dplyr::glimpse(typical_data)
+
+glimpse(typical_data)
 
 
 ## ----load-data-----------------------------------------------------------
@@ -30,8 +31,6 @@ vis_miss(typical_data)
 
 
 ## ----vis_dat-------------------------------------------------------------
-
-library(visdat)
 
 vis_dat(airquality)
 
@@ -79,11 +78,87 @@ vis_miss(airquality,
          cluster = TRUE)
 
 
-## ----plotly-example------------------------------------------------------
+## ----vis-compare-iris----------------------------------------------------
+chickwts_diff <- chickwts
+chickwts_diff[sample(1:nrow(chickwts), 30),sample(1:ncol(chickwts), 2)] <- NA
 
-library(plotly)
+vis_compare(chickwts_diff, chickwts)
 
-vis_dat(airquality) %>% ggplotly()
-vis_miss(airquality) %>% ggplotly()
 
+## ----vis-compare-error, eval = FALSE-------------------------------------
+#  
+#  chickwts_diff_2 <- chickwts
+#  chickwts_diff_2$new_col <- chickwts_diff_2$weight*2
+#  
+#  vis_compare(chickwts, chickwts_diff_2)
+#  # Error in vis_compare(chickwts, chickwts_diff_2) :
+#  #   Dimensions of df1 and df2 are not the same. vis_compare requires dataframes of identical dimensions.
+
+## ----vis-expect----------------------------------------------------------
+
+vis_expect(airquality, ~.x >= 25)
+
+
+## ----vis-expect-bad-strings----------------------------------------------
+
+bad_data <- data.frame(x = c(rnorm(100), rep("N/A", 10)),
+                       y = c(rep("N A ", 30), rnorm(80)))
+
+vis_expect(bad_data, ~.x %in% c("N/A", "N A "))
+
+## ----vis-cor-------------------------------------------------------------
+
+vis_cor(airquality)
+
+
+## ----vis-cor-spearman----------------------------------------------------
+
+vis_cor(airquality, cor_method = "spearman")
+
+
+## ----vis-cor-na-action---------------------------------------------------
+
+vis_cor(airquality,
+        na_action = "complete.obs")
+
+
+## ----create-messy-vec----------------------------------------------------
+
+messy_vector <- c(TRUE,
+                  T,
+                  "TRUE",
+                  "T",
+                  "01/01/01",
+                  "01/01/2001",
+                  NA,
+                  NaN,
+                  "NA",
+                  "Na",
+                  "na",
+                  "10",
+                  10,
+                  "10.1",
+                  10.1,
+                  "abc",
+                  "$%TG")
+
+set.seed(1114)
+messy_df <- data.frame(var1 = messy_vector,
+                       var2 = sample(messy_vector),
+                       var3 = sample(messy_vector))
+
+
+## ----vis-guess-messy-df, fig.show='hold', out.width='50%'----------------
+
+vis_guess(messy_df)
+vis_dat(messy_df)
+
+
+## ----intx, eval = FALSE--------------------------------------------------
+#  
+#  library(plotly)
+#  ggplotly(vis_dat(airquality))
+#  ggplotly(vis_miss(airquality))
+#  ggplotly(vis_guess(airquality))
+#