+++ tests/testthat/test-vis-large.R
@@ -1,8 +1,8 @@
 context("Warn for large data")
 
 big_df <- matrix(rnorm(100000),
-       nrow = 1000,
-       ncol = 1000) %>%
+                 nrow = 1000,
+                 ncol = 1000) %>%
   as.data.frame()
 
 test_that("vis_dat and vis_miss throw warnings when the DF is above size",{
@@ -10,6 +10,3 @@ test_that("vis_dat and vis_miss throw warnings when the DF is above size",{
   expect_error(vis_dat(big_df))
   expect_error(vis_miss(big_df))
 })
-
-
-