+++ tests/testthat/test-visdat-internals.R
@@ -3,11 +3,11 @@ context("Internals")
 test_vis_gather_ <- suppressWarnings(vis_gather_(typical_data))
 
 suppressWarnings(
-test_old_gather <- typical_data %>%
-  dplyr::mutate(rows = seq_len(nrow(.))) %>%
-  tidyr::gather_(key_col = "variable",
-                 value_col = "valueType",
-                 gather_cols = names(.)[-length(.)])
+  test_old_gather <- typical_data %>%
+    dplyr::mutate(rows = seq_len(nrow(.))) %>%
+    tidyr::gather_(key_col = "variable",
+                   value_col = "valueType",
+                   gather_cols = names(.)[-length(.)])
 )
 
 test_that("vis_gather_ returns the same as previous",{
@@ -22,9 +22,9 @@ d_old <- typical_data %>%
   vis_gather_()
 
 suppressWarnings(
-d_old$value <-  tidyr::gather_(typical_data,
-                               "variables",
-                               "value", names(typical_data))$value
+  d_old$value <-  tidyr::gather_(typical_data,
+                                 "variables",
+                                 "value", names(typical_data))$value
 )
 
 d_new <-
@@ -36,4 +36,11 @@ d_new <-
 
 test_that("vis_extract_value performs the same as old method",{
   expect_equal(d_old,d_new)
+})
+
+test_that("any_numeric returns TRUE for numeric dataframes and FALSE for dataframes containing non-numeric values",{
+
+  testthat::expect_equal(all_numeric(airquality),TRUE)
+  testthat::expect_equal(all_numeric(iris),FALSE)
+
 })